More CDI handling of apps

+++ testsuite/testsuite-microprofile-metrics/src/test/java/org/wildfly/swarm/microprofile_metrics/MicroprofileMetricsTest.java
@@ -446,9 +446,9 @@ public class MicroprofileMetricsTest {
 
                 .body("'metricTest.test1.countMeA'", equalTo(1))
 
-                .body("'metricTest.test1.gauge'", equalTo(19))
+                .body("'metricTest.test1.gauge'", equalTo((float)19)) // TODO should be int
 
-                .body("'org.wildfly.swarm.microprofile_metrics.MetricAppBean.gaugeMeA'", equalTo(1000))
+                .body("'org.wildfly.swarm.microprofile_metrics.MetricAppBean.gaugeMeA'", equalTo((float)1000)) // TODO should be int
 
                 .body("'metricTest.test1.histogram'.count", equalTo(1000))
                 .body("'metricTest.test1.histogram'.max", equalTo(999))
